src/PythonBackend/requirements.txt
fastapi==0.112.2
uvicorn
python-dotenv
psycopg2-binary